Fixed, yo

+++ Take 2/Assets/Scripts/Player/PlayerBulletPool.cs
@@ -22,7 +22,7 @@ public class PlayerBulletPool : MonoBehaviour {
     }
 
     //Start position
-    public void ActivateBullet(Vector2 startPos, float velocity, Vector2 direction, float damage)
+    public void ActivateBullet(Vector2 startPos, float velocity, Vector2 direction, uint damage)
     {
         foreach (GameObject bullet in bulletList)
         {
